% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisdom_on_startup.R
\name{wisdom_on_startup}
\alias{wisdom_on_startup}
\title{wisdom_on_startup}
\usage{
wisdom_on_startup()
}
\description{
Dispenses a randomly selected piece of wisdom from Merlin Mann's \href{https://github.com/merlinmann/wisdom}{Wisdom Project} on R startup.
}
